+++ config.go
@@ -9,8 +9,9 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/hashicorp/go-plugin"
 	"github.com/hashicorp/hcl"
-	"github.com/hashicorp/terraform/plugin"
+	tfplugin "github.com/hashicorp/terraform/plugin"
 	"github.com/hashicorp/terraform/terraform"
 	"github.com/kardianos/osext"
 )
@@ -202,7 +203,9 @@ func (c *Config) providerFactory(path string) terraform.ResourceProviderFactory 
 	// Build the plugin client configuration and init the plugin
 	var config plugin.ClientConfig
 	config.Cmd = pluginCmd(path)
+	config.HandshakeConfig = tfplugin.Handshake
 	config.Managed = true
+	config.Plugins = tfplugin.PluginMap
 	client := plugin.NewClient(&config)
 
 	return func() (terraform.ResourceProvider, error) {
@@ -213,7 +216,12 @@ func (c *Config) providerFactory(path string) terraform.ResourceProviderFactory 
 			return nil, err
 		}
 
-		return rpcClient.ResourceProvider()
+		raw, err := rpcClient.Dispense(tfplugin.ProviderPluginName)
+		if err != nil {
+			return nil, err
+		}
+
+		return raw.(terraform.ResourceProvider), nil
 	}
 }
 
@@ -232,8 +240,10 @@ func (c *Config) ProvisionerFactories() map[string]terraform.ResourceProvisioner
 func (c *Config) provisionerFactory(path string) terraform.ResourceProvisionerFactory {
 	// Build the plugin client configuration and init the plugin
 	var config plugin.ClientConfig
+	config.HandshakeConfig = tfplugin.Handshake
 	config.Cmd = pluginCmd(path)
 	config.Managed = true
+	config.Plugins = tfplugin.PluginMap
 	client := plugin.NewClient(&config)
 
 	return func() (terraform.ResourceProvisioner, error) {
@@ -242,7 +252,12 @@ func (c *Config) provisionerFactory(path string) terraform.ResourceProvisionerFa
 			return nil, err
 		}
 
-		return rpcClient.ResourceProvisioner()
+		raw, err := rpcClient.Dispense(tfplugin.ProvisionerPluginName)
+		if err != nil {
+			return nil, err
+		}
+
+		return raw.(terraform.ResourceProvisioner), nil
 	}
 }
